Wait complete

+++ 3rd_party/darknet/Network.cpp
@@ -135,7 +135,7 @@ void Network::forward() {
 	if (use_gpu) {
 		cuda_push_array(input_gpu, input, inputs*batch);
 
-		for (int i = 0; i < layers.size(); ++i) {
+		for (size_t i = 0; i < layers.size(); ++i) {
 			auto& l = layers[i];
 			if (l->delta) {
 				fill_cpu(l->outputs * l->batch, 0, l->delta, 1);
